cart design modified

+++ src/Redux/Reducers/cartReducer.js
@@ -2,15 +2,26 @@ function cartReducer(cart = JSON.parse(localStorage.getItem('carts')) || [], act
 	switch (action.type) {
 		case 'ADD_TO_CART': {
 			const existingCart = [...cart];
-			const newItem = action.payload;
-			const exists = existingCart.find(product => product.id === newItem.id);
-            if (exists) {
+			const newItemToAdd = action.payload;
+            const minBuyQuanity = newItemToAdd.minBuyQuanity;
+            const maxBuyQuantity = newItemToAdd.maxBuyQuantity;
+
+            console.log('newItemToAdd: ',newItemToAdd);
+            console.log('minBuyQuanity: ',minBuyQuanity);
+            console.log('maxBuyQuantity: ',maxBuyQuantity);
+
+			const exists = existingCart.find(product => product.id === newItemToAdd.id);
+
+            if (exists && (exists.cartQuantity < maxBuyQuantity)) {
 				exists.cartQuantity = exists.cartQuantity + 1;
 				localStorage.setItem('carts', JSON.stringify(existingCart));
                 return existingCart;
 			}
+            else if(exists && (exists.cartQuantity = maxBuyQuantity)){
+                return existingCart;
+            }
              else {
-				action.payload.cartQuantity = 1;
+				action.payload.cartQuantity = minBuyQuanity;
 				const newCart = [...existingCart, action.payload];
 				localStorage.setItem('carts', JSON.stringify(newCart));
                 return newCart;
@@ -47,8 +58,8 @@ export default cartReducer;
 //         const existingCart = [...localItems];
 //         // console.log('existingCart', existingCart);
 //         // console.log('current state', state.cartItems);
-//         const newItem = action.payload;
-//         const itemExist = existingCart.find((item) => item.id === newItem.id);
+//         const newItemToAdd = action.payload;
+//         const itemExist = existingCart.find((item) => item.id === newItemToAdd.id);
   
 //         // console.log('itemExist', itemExist);
 //         // console.log('current state', state.cartItems);
